+++ src/main/java/upbrella/be/rent/repository/RentRepositoryCustom.java
@@ -1,7 +1,7 @@
 package upbrella.be.rent.repository;
 
 import org.springframework.data.domain.Pageable;
-import upbrella.be.rent.dto.HistoryInfoDto;
+import upbrella.be.rent.dto.response.HistoryInfoDto;
 import upbrella.be.rent.dto.request.HistoryFilterRequest;
 import upbrella.be.rent.entity.History;
